pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["dvsmith", "dvsmith.cli", "dvsmith.config", "dvsmith.cvdp", "dvsmith.core", "dvsmith.adapters", "dvsmith.adapters.sim", "dvsmith.adapters.cov", "dvsmith.adapters.parse", "dvsmith.harness"]

[project]
name = "dv-smith"
version = "0.2.0"
description = "Convert SystemVerilog/UVM testbenches into DV gyms"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["verification", "uvm", "systemverilog", "eda", "testing", "gym"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Testing",
    "Topic :: Scientific/Engineering :: Electronic Design Automation (EDA)",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

dependencies = [
    "pyyaml>=6.0",
    "claude-agent-sdk>=0.1.0",
    "python-dotenv>=1.0.0",
    "anthropic>=0.18.0",
    "tqdm>=4.65.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "typer>=0.9.0",
    "rich>=13.0.0",
    "tenacity>=8.0.0",
    "filelock>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "mypy>=1.0",
    "flake8>=6.0",
]

[project.scripts]
dvsmith = "dvsmith.cli.app:main"
dvsmith-old = "dvsmith.cli_old:main"

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]